+++ nengo/networks/tests/test_ensemblearray.py
@@ -4,7 +4,6 @@ import pytest
 import nengo
 from nengo.dists import Choice
 from nengo.utils.compat import range
-from nengo.utils.testing import WarningCatcher
 
 
 def test_multidim(Simulator, plt, seed, rng):
@@ -202,27 +201,39 @@ def test_arguments():
         nengo.networks.EnsembleArray(nengo.LIF(10), 1, dimensions=2)
 
 
-def test_neuronconnection(Simulator, nl, seed):
-    catcher = WarningCatcher()
+def test_directmode_errors():
+    with nengo.Network() as net:
+        net.config[nengo.Ensemble].neuron_type = nengo.Direct()
+
+        ea = nengo.networks.EnsembleArray(10, 2)
+        with pytest.raises(TypeError):
+            ea.add_neuron_input()
+        with pytest.raises(TypeError):
+            ea.add_neuron_output()
+
+
+def test_neuroninput(Simulator, seed):
     with nengo.Network(seed=seed) as net:
-        net.config[nengo.Ensemble].neuron_type = nl()
+        inp = nengo.Node([-10] * 20)
+        ea = nengo.networks.EnsembleArray(10, 2)
+        ea.add_neuron_input()
+        nengo.Connection(inp, ea.neuron_input, synapse=None)
+        p = nengo.Probe(ea.output)
 
-        input = nengo.Node([-10] * 20)
-        with catcher:
-            ea = nengo.networks.EnsembleArray(10, 2, neuron_nodes=True)
+    s = Simulator(net)
+    s.run(0.01)
+    assert np.all(s.data[p] < 1e-2) and np.all(s.data[p] > -1e-2)
 
-        nengo.Connection(input, ea.neuron_input)
 
+def test_neuronoutput(Simulator, seed):
+    with nengo.Network(seed=seed) as net:
+        ea = nengo.networks.EnsembleArray(
+            10, 2, encoders=nengo.dists.Choice([[1]]))
+        ea.add_neuron_output()
+        inp = nengo.Node([-10, -10])
+        nengo.Connection(inp, ea.input, synapse=None)
         p = nengo.Probe(ea.neuron_output)
 
     s = Simulator(net)
-    s.run(0.1)
-
-    # Some nls (e.g. Sigmoid) never go all the way to 0
-    assert np.all(s.data[p][-1] < 1e-2)
-
-    if nl == nengo.Direct:
-        assert (len(catcher.record) == 1 and
-                catcher.record[0].category is UserWarning)
-    else:
-        assert len(catcher.record) == 0
+    s.run(0.01)
+    assert np.all(s.data[p] < 1e-2)