+++ nengo/networks/workingmemory.py
@@ -18,15 +18,13 @@ def InputGatedMemory(n_neurons, dimensions, feedback=1.0,
 
     with net:
         # integrator to store value
-        net.mem = EnsembleArray(n_neurons, dimensions,
-                                neuron_nodes=True, label="mem")
+        net.mem = EnsembleArray(n_neurons, dimensions, label="mem")
         nengo.Connection(net.mem.output, net.mem.input,
                          transform=feedback,
                          synapse=recurrent_synapse)
 
         # calculate difference between stored value and input
-        net.diff = EnsembleArray(n_neurons, dimensions,
-                                 neuron_nodes=True, label="diff")
+        net.diff = EnsembleArray(n_neurons, dimensions, label="diff")
         nengo.Connection(net.mem.output, net.diff.input, transform=-1)
 
         # feed difference into integrator
@@ -37,13 +35,14 @@ def InputGatedMemory(n_neurons, dimensions, feedback=1.0,
         # gate difference (if gate==0, update stored value,
         # otherwise retain stored value)
         net.gate = nengo.Node(size_in=1)
+        net.diff.add_neuron_input()
         nengo.Connection(net.gate, net.diff.neuron_input,
                          transform=np.ones((n_total_neurons, 1)) * -10,
                          synapse=None)
 
         # reset input (if reset=1, remove all values, and set to 0)
         net.reset = nengo.Node(size_in=1)
-        nengo.Connection(net.reset, net.mem.neuron_input,
+        nengo.Connection(net.reset, net.mem.add_neuron_input(),
                          transform=np.ones((n_total_neurons, 1)) * -3,
                          synapse=None)